requirements.txt
python-dotenv==1.0.1
openai
langchain==0.1.19
langchain-openai
moviepy